Added api/profile/ see, like, unlike, ban, unban, fake. Modified setup.sql and api/profile/get. Some fix.

+++ server/controllers/ImageController/deletePhoto.js
@@ -1,9 +1,17 @@
 const fs = require('fs');
 
 module.exports = (req, res) => {
+    if (req.session.id === undefined) {
+        res.send('Need login');
+        return;
+    }
+    let {name} = req.body;
+    if (!name) {
+        res.send('Need Photo Name');
+        return;
+    }
     let db      = req.app.get('db'),
         rootDir = req.app.get('rootDir'),
-        {name}  = req.body,
         promise = db.delete('photos', ['filename', 'user_id'], [name, req.session.id]),
         error   = (e) => {
             console.log('error: ', e);
